Fixed last comments again haha (3 hours) Just added image for number line, hope it sticks
i know i disabled the buttons in the subtract, so i don't know what they are going on about.
just removed them when not needed (probably will be asked to change it)

Added new images with x and 1 inside

removed tablet layout might be the reason for the problems encountered by daniel

some values have been hard coded in layouts, if there are problems, might be there. (5dp etc...)